# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_javinizer_javinizer-go_internal_api_contracts.PathAutocompleteSuggestion:
    properties:
      is_dir:
        example: true
        type: boolean
      name:
        example: videos
        type: string
      path:
        example: /path/to/videos
        type: string
    type: object
...
  internal_api_events.ErrorResponse:
    properties:
      error:
        example: Movie not found
        type: string
      errors:
        items:
          type: string
        type: array
    type: object
...
  internal_api_events.eventStatsResponse:
    properties:
      by_severity:
        additionalProperties:
          format: int64
          type: integer
        type: object
      by_source:
        additionalProperties:
          format: int64
          type: integer
        type: object
      by_type:
        additionalProperties:
          format: int64
          type: integer
        type: object
      total:
        type: integer
    type: object
...
  internal_api_file.ErrorResponse:
    properties:
      error:
        example: Movie not found
        type: string
      errors:
        items:
          type: string
        type: array
    type: object
  internal_api_file.PathAutocompleteRequest:
    properties:
      limit:
        example: 10
        type: integer
      path:
        example: /path/to/vid
        type: string
    required:
    - path
    type: object
  internal_api_file.PathAutocompleteResponse:
    properties:
      base_path:
        example: /path/to
        type: string
      input_path:
        example: /path/to/vid
        type: string
      suggestions:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.PathAutocompleteSuggestion'
        type: array
    type: object
...
  internal_api_history.DeleteHistoryBulkResponse:
    properties:
      deleted:
        type: integer
    type: object
  internal_api_history.ErrorResponse:
    properties:
      error:
        example: Movie not found
        type: string
      errors:
        items:
          type: string
        type: array
    type: object
  internal_api_history.HistoryListResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      records:
        items:
          $ref: '#/definitions/internal_api_history.HistoryRecord'
        type: array
      total:
        type: integer
    type: object
  internal_api_history.HistoryRecord:
    properties:
      created_at:
        type: string
      dry_run:
        type: boolean
      error_message:
        type: string
      id:
        type: integer
      metadata:
        type: string
      movie_id:
        type: string
      new_path:
        type: string
      operation:
        type: string
      original_path:
        type: string
      status:
        type: string
    type: object
...
  internal_api_jobs.ErrorResponse:
    properties:
      error:
        example: Movie not found
        type: string
      errors:
        items:
          type: string
        type: array
    type: object
  internal_api_jobs.JobListItem:
    properties:
      completed:
        example: 9
        type: integer
      completed_at:
        example: "2026-04-12T10:05:00Z"
        type: string
      destination:
        example: /path/to/output
        type: string
      failed:
        example: 1
        type: integer
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      operation_count:
        example: 10
        type: integer
      organized_at:
        example: "2026-04-12T10:05:00Z"
        type: string
      progress:
        example: 0.9
        type: number
      reverted_at:
        example: "2026-04-12T11:00:00Z"
        type: string
      reverted_count:
        example: 7
        type: integer
      started_at:
        example: "2026-04-12T10:00:00Z"
        type: string
      status:
        example: organized
        type: string
      total_files:
        example: 10
        type: integer
    type: object
...
  internal_api_token.ErrorResponse:
    properties:
      error:
        example: Movie not found
        type: string
      errors:
        items:
          type: string
        type: array
    type: object
  internal_api_token.createTokenRequest:
    properties:
      name:
        type: string
    type: object
  internal_api_token.tokenListItem:
    properties:
      created_at:
        type: string
      id:
        type: string
      last_used_at:
        type: string
      name:
        type: string
      token_prefix:
        type: string
    type: object
  internal_api_token.tokenListResponse:
    properties:
      count:
        type: integer
      tokens:
        items:
          $ref: '#/definitions/internal_api_token.tokenListItem'
        type: array
    type: object
  internal_api_token.tokenResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      last_used_at:
        type: string
      name:
        type: string
      token:
        type: string
      token_prefix:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/fedora-oss/javinizer-go/issues
  description: REST API for JAV metadata scraping and file organization
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/fedora-oss/javinizer-go
  title: Javinizer API
  version: "1.0"
paths:
...
  /api/v1/browse/autocomplete:
    post:
      consumes:
      - application/json
      description: Returns directory suggestions for a partially typed path
      parameters:
      - description: Autocomplete parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_file.PathAutocompleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_file.PathAutocompleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_file.ErrorResponse'
      summary: Autocomplete directory path
      tags:
      - web
...
  /api/v1/events/stats:
    get:
      description: Get event counts grouped by type, severity, and source
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_events.eventStatsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_events.ErrorResponse'
      summary: Get event statistics
      tags:
      - events
...
  /api/v1/history:
    delete:
      description: Delete multiple history records based on criteria
      parameters:
      - description: Delete records older than N days
        in: query
        name: older_than_days
        type: integer
      - description: Delete all records for a specific movie
        in: query
        name: movie_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_history.DeleteHistoryBulkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_history.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_history.ErrorResponse'
      summary: Delete history records in bulk
      tags:
      - history
    get:
      description: Get a paginated list of history records with optional filtering
      parameters:
      - description: 'Number of records to return (default: 50, max: 500)'
        in: query
        name: limit
        type: integer
      - description: 'Number of records to skip (default: 0)'
        in: query
        name: offset
        type: integer
      - description: Filter by operation type (scrape, organize, download, nfo)
        in: query
        name: operation
        type: string
      - description: Filter by status (success, failed, reverted)
        in: query
        name: status
        type: string
      - description: Filter by movie ID
        in: query
        name: movie_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_history.HistoryListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_history.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_history.ErrorResponse'
      summary: List history records
      tags:
      - history
...
  /api/v1/jobs/{id}:
    get:
      description: Get details for a specific batch job by ID
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_jobs.JobListItem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_jobs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_jobs.ErrorResponse'
      summary: Get a single batch job
      tags:
      - jobs
...
  /api/v1/tokens:
    get:
      description: List all active (non-revoked) API tokens
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_token.tokenListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_token.ErrorResponse'
      summary: List API tokens
      tags:
      - tokens
    post:
      consumes:
      - application/json
      description: Create a new API token with an optional name
      parameters:
      - description: Token details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_token.createTokenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api_token.tokenResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_token.ErrorResponse'
      summary: Create API token
      tags:
      - tokens
  /api/v1/tokens/{id}:
    delete:
      description: Revoke an API token by ID, making it immediately invalid
      parameters:
      - description: Token ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_token.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_token.ErrorResponse'
      summary: Revoke API token
      tags:
      - tokens
  /api/v1/tokens/{id}/regenerate:
    post:
      description: Regenerate an API token by ID, returning a new token value and
        invalidating the old one
      parameters:
      - description: Token ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_token.tokenResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_token.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_token.ErrorResponse'
      summary: Regenerate API token
      tags:
      - tokens
...
schemes:
- http
- https
swagger: "2.0"